requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0.1
django-filter==23.5
django-model-utils==4.3.1
django-notifications-hq==1.8.3
djangorestframework==3.14.0
idna==3.6
jsonfield==3.1.0
Markdown==3.5.1
pillow==10.2.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
swapper==1.3.0
tzdata==2023.4
urllib3==2.1.0